licitometro/backend/recon/requirements.txt
scrapy==2.8.0
sqlalchemy==2.0.15
celery==5.2.7
redis==4.5.4
psycopg2-binary==2.9.6
requests==2.30.0
logging==0.5.1.2
typing==3.7.4.3
sqlalchemy-utils==0.41.1
python-dotenv==1.0.0
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
python-dateutil==2.8.2
flower==1.2.0  # Monitoreo de tareas Celery
